fixed issue with mutable objects in factory fixtures and moved them to conftest.py
